% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitbitAnalysis-package.R
\docType{package}
\name{FitbitAnalysis-package}
\alias{FitbitAnalysis}
\alias{FitbitAnalysis-package}
\title{FitbitAnalysis: Exploratory Analysis of Fitbit Sensor Data}
\description{
Ultra-basic set of R functions for importing a Fitbit archive and running exploratory analyses on top.
}
\author{
\strong{Maintainer}: Kay H. Brodersen \email{khbrodersen@gmail.com} (\href{https://orcid.org/0000-0002-7707-090X}{ORCID})

}
\keyword{internal}
