% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{LoadFitbitData}
\alias{LoadFitbitData}
\title{Load a Fitbit archive from disk}
\usage{
LoadFitbitData(input_path)
}
\arguments{
\item{input_path}{String containing the path of the extracted Fitbit archive.
The string should end on a 'FirstnameLastname' directory.}
}
\value{
A data frame keyed on \code{date} and one column per metric.
}
\description{
Load a Fitbit archive, as downloaded from fitbit.com, Settings,
Data Export, Request Data. The downloaded zip file must be extracted into a
folder. See below for an example.
}
\examples{
\dontrun{
data <- LoadFitbitData("/path/to/MyFitbitData/FirstnameLastname/")
head(data)
PlotMetrics(data)
PlotMetrics(data \%>\% dplyr::select(date, VeryActiveMinutes))
PlotWeeklyMetrics(data)
PlotWeekdayMetrics(data)
PlotTimeAtRest(data)
PlotMinutesAwakeVsStartHour(data)
}
}
