% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{PlotMetrics}
\alias{PlotMetrics}
\title{Plot daily values of one or more metrics over time}
\usage{
PlotMetrics(data)
}
\arguments{
\item{data}{Data frame keyed on \code{date}, with one or more metrics columns
containing numeric values.}
}
\description{
Plot daily values of one or more metrics over time
}
