% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{PlotMinutesAwakeVsStartHour}
\alias{PlotMinutesAwakeVsStartHour}
\title{Plot number of minutes awake vs. start time of time at rest}
\usage{
PlotMinutesAwakeVsStartHour(data, max_start_hour = Inf)
}
\arguments{
\item{data}{Data frame of: date, StartTime, MinutesAsleep.}

\item{max_start_hour}{Maximum start hour to clip plot at (e.g., 28).}
}
\description{
Plot number of minutes awake vs. start time of time at rest
}
