% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{PlotTimeAtRest}
\alias{PlotTimeAtRest}
\title{Plot times at rest}
\usage{
PlotTimeAtRest(data, histogram_bins = 50)
}
\arguments{
\item{data}{Data frame of: date, StartTime, EndTime.}

\item{histogram_bins}{Number of histogram bins for plotting.}
}
\description{
Plot times at rest
}
