% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{PlotTimeAtRestCDF}
\alias{PlotTimeAtRestCDF}
\title{Plot a CDF of times at rest}
\usage{
PlotTimeAtRestCDF(data)
}
\arguments{
\item{data}{Data frame of: date, StartTime, EndTime, MinutesAsleep.}
}
\description{
Plot a CDF of times at rest
}
