% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{PlotWeekdayMetrics}
\alias{PlotWeekdayMetrics}
\title{Plot day-of-week averages of one or more metrics}
\usage{
PlotWeekdayMetrics(data)
}
\arguments{
\item{data}{Data frame keyed on \code{date}, with one or more metrics columns
containing numeric values.}
}
\description{
Plot day-of-week averages of one or more metrics
}
