% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{PlotWeeklyMetrics}
\alias{PlotWeeklyMetrics}
\title{Plot weekly sums of one or more metrics over time}
\usage{
PlotWeeklyMetrics(data)
}
\arguments{
\item{data}{Data frame keyed on \code{date}, with one or more metrics columns
containing numeric values.}
}
\description{
Plot weekly sums of one or more metrics over time
}
