% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{PlotMinutesAwakeVsStartHour}
\alias{PlotMinutesAwakeVsStartHour}
\title{Plot number of minutes awake vs. start time of time at rest}
\usage{
PlotMinutesAwakeVsStartHour(data)
}
\arguments{
\item{data}{Data frame of: date, StartTime, MinutesAsleep.}
}
\description{
Plot number of minutes awake vs. start time of time at rest
}
